Changes as suggested in review
