+++ lib/routes/prototype-admin-routes.js
@@ -10,10 +10,8 @@ const { projectDir, packageDir } = require('../path-utils')
 const path = require('path')
 const extensions = require('../extensions/extensions')
 const config = require('../config').getConfig()
-const { requestHttpsJson } = require('../utils')
+const { requestHttpsJson, scripts } = require('../utils')
 const { exec } = require('../exec')
-
-let extensionConfig = {}
 
 const appViews = extensions.getAppViews([
   path.join(projectDir, 'node_modules'),
@@ -232,7 +230,7 @@ function locateTemplateConfig (req) {
 
 router.get('/templates/view', (req, res) => {
   const model = {
-    extensionConfig,
+    extensionConfig: extensions.getAppConfig({ scripts }),
     serviceName: 'Service name goes here'
   }
   const templateConfig = locateTemplateConfig(req)
@@ -486,7 +484,4 @@ router.post('/plugins/uninstall', async (req, res) => {
   }
 })
 
-module.exports = (scripts) => {
-  extensionConfig = extensions.getAppConfig({ scripts })
-  return router
-}
+module.exports = router
