+++ lib/utils.js
@@ -14,7 +14,8 @@ const config = require('./config').getConfig()
 const extensions = require('./extensions/extensions')
 const filters = require('./filters/api')
 const routes = require('./routes/api')
-const { appDir, tempDir } = require('./path-utils')
+const { appDir, tempDir, projectDir } = require('./path-utils')
+const { existsSync } = require('fs')
 
 // Tweak the Markdown renderer
 const defaultMarkedRenderer = marked.defaults.renderer || new marked.Renderer()
@@ -29,6 +30,15 @@ marked.use({
     }
   }
 })
+
+const scripts = []
+if (existsSync(path.join(projectDir, 'app', 'assets', 'javascripts', 'application.js'))) {
+  scripts.push('/public/javascripts/application.js')
+}
+if (extensions.legacyGovukFrontendFixesNeeded()) {
+  scripts.push('/extension-assets/govuk-prototype-kit/lib/assets/javascripts/optional/legacy-govuk-frontend-init.js')
+}
+exports.scripts = scripts
 
 // Require core and custom filters, merges to one object
 // and then add the methods to Nunjucks environment
